requirements.txt
pybullet
numpy
paho-mqtt
matplotlib
networkx